rename package: common.shared -> common.domain
